Move library to root of repository
This structure is required in order to allow the library to be added to the Arduino Library Manager index.

Another benefit is to facilitate use with the popular alternative library installation technique of downloading the .zip files of the library GitHub generates (Clone or download > Download ZIP or the Releases page) and then using the Arduino IDE's Sketch > Include Library > Add .ZIP Library on the downloaded file. This requires the library to be in the root of the repository, not in a subfolder. If the library is not in the root of the repository this installation technique fails:

Specified folder/zip file does not contain a valid library

This is the standard repository structure used in all official Arduino libraries:
https://github.com/arduino-libraries